refactor(auth): combined auth/login modules
