+++ src/features/auth/types.ts
@@ -1,0 +1,6 @@
+export type UserType = {
+  id: string
+  firstName: string
+  lastName: string
+  email: string
+}
