+++ src/features/events/components/EventsList/_stories/EventsListE2E.stories.tsx
@@ -2,7 +2,7 @@ import type { Story, Meta } from '@storybook/react/types-6-0'
 import React from 'react'
 import { QueryClient, QueryClientProvider } from 'react-query'
 
-import { UserContextProvider } from '~/features/auth/contexts/userContext'
+import { UserContextProvider } from '~/features/auth/contexts/user'
 import { EventsList } from '~/features/events/components/EventsList'
 import { EventFilterContextProvider } from '~/features/events/contexts/event-filter'
 import { EventViewContextProvider } from '~/features/events/contexts/event-view'
