--- src/features/login/index.ts
@@ -1,1 +1,0 @@
-export * from './pages/LoginPage'
